split buffer into several files
