+++ tianshou/data/__init__.py
@@ -1,15 +1,13 @@
 from tianshou.data.batch import Batch
 from tianshou.data.utils.converter import to_numpy, to_torch, to_torch_as
 from tianshou.data.utils.segtree import SegmentTree
-from tianshou.data.buffer import (
-    ReplayBuffer,
-    PrioritizedReplayBuffer,
-    ReplayBufferManager,
-    PrioritizedReplayBufferManager,
-    VectorReplayBuffer,
-    PrioritizedVectorReplayBuffer,
-    CachedReplayBuffer,
-)
+from tianshou.data.buffer.base import ReplayBuffer
+from tianshou.data.buffer.prio import PrioritizedReplayBuffer
+from tianshou.data.buffer.manager import ReplayBufferManager
+from tianshou.data.buffer.manager import PrioritizedReplayBufferManager
+from tianshou.data.buffer.vecbuf import VectorReplayBuffer
+from tianshou.data.buffer.vecbuf import PrioritizedVectorReplayBuffer
+from tianshou.data.buffer.cached import CachedReplayBuffer
 from tianshou.data.collector import Collector, AsyncCollector
 
 __all__ = [
